Add StringView::toExistingAtomicString() https://bugs.webkit.org/show_bug.cgi?id=171405
Reviewed by Andreas Kling.

Similar to the reasons for JSString::toExistingAtomicString() we should expose a way to
convert a StringView to an existing atomic string. Looking up an atomic string is more
efficient that creating one, which requires a lookup and memory allocation.

We are not making use of StringView::toExistingAtomicString() now, but will in the patch
for <https://bugs.webkit.org/show_bug.cgi?id=170925>.

* wtf/text/AtomicStringImpl.cpp:
(WTF::AtomicStringImpl::lookUp): Modified to take a const LChar*/UChar*.
(WTF::AtomicStringImpl::lookUpInternal): Renamed to AtomicStringImpl::lookup() to avoid
an extra function call.
* wtf/text/AtomicStringImpl.h:
* wtf/text/StringView.h:
(WTF::StringView::toExistingAtomicString): Added.

Identifier: 188340@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215947 268f45cc-cd09-0410-ab3c-d52691b4dbfc