App crashing: Dispatch queue: com.apple.root.user-interactive-qos / vBoxConvolve / WebCore::FEGaussianBlur::platformApplySoftware() https://bugs.webkit.org/show_bug.cgi?id=171461 <rdar://problem/30534722>
Reviewed by Eric Carlson.

Source/WebCore:

We're getting reports of crashes in this function, caused by null or empty data being
passed to vImage. Guard against this, in a way that will ASSERT in debug builds if
anyone comes across it.

Test: css3/filters/blur-various-radii.html

* platform/graphics/filters/FEGaussianBlur.cpp:
(WebCore::accelerateBoxBlur): Return early if things don't look good.

LayoutTests:

Test a bunch of blurs a frame at a time.

* css3/filters/blur-various-radii-expected.html: Added.
* css3/filters/blur-various-radii.html: Added.

Identifier: 188350@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215957 268f45cc-cd09-0410-ab3c-d52691b4dbfc