iBooks text can overlap, sometimes columns are shifted splitting words. https://bugs.webkit.org/show_bug.cgi?id=171472 <rdar://problem/31096037>
Reviewed by Antti Koivisto.

Source/WebCore:

Instead of just checking if the glyph is taller than the line, we need to ensure that both the
ascent and the descent have enough space (this is for -webkit-line-box-contain: glyph).

Test: fast/text/simple-line-layout-glyph-overflows-line.html

* rendering/SimpleLineLayout.cpp:
(WebCore::SimpleLineLayout::canUseForText): compute the available space for the ascent/descent
and check them against the ceil-ed(see FontCascade::floatWidthForSimpleText) glyph min/max y.

LayoutTests:

* fast/text/simple-line-layout-glyph-overflows-line-expected.html: Added.
* fast/text/simple-line-layout-glyph-overflows-line.html: Added.

Identifier: 188367@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215976 268f45cc-cd09-0410-ab3c-d52691b4dbfc