Different behaviour with the .sort(callback) method (unlike Firefox & Chrome) https://bugs.webkit.org/show_bug.cgi?id=47825
Reviewed by Saam Barati.

JSTests:

* stress/sorting-boolean-result-comparator.js: Added.
(checkArray):

Source/JavaScriptCore:

This patch makes our sort function match the behavior of Firefox
and Chrome when the result of the comparison function is a
boolean. When we first switched to using merge sort, it regressed
JQuery sorting of DOM nodes by 30%. The regression was do to the
fact that JQuery was using compareDocumentPosition to compare the
locations of objects. Since one of the benchmarks would pass a
reverse sorted list to the sort function we would end up walking
the entire DOM to do comparisons. The solution to this was to
merge based on comparison(right, left) rather than
comparison(left, right). Although, in practice this does nothing
since sort could just as easily receive an already sorted list and
we're back in the same spot.

The downside of sorting with comparison(right, left) is that to
maintain stability when sorting, you only want to merge from right
when the comparison function returns a negative value. This is
where the problem with booleans comes in. Since booleans toNumber
false to 0 and true to 1 both values are "equal". This patch fixes
this by special casing boolean return values.

* builtins/ArrayPrototype.js:
(sort.merge):

LayoutTests:

Fix broken test.

* http/tests/inspector/worker/blob-script-with-cross-domain-imported-scripts-expected.txt:


Identifier: 188523@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216137 268f45cc-cd09-0410-ab3c-d52691b4dbfc