AX: Propagate aria-readonly to grid descendants https://bugs.webkit.org/show_bug.cgi?id=171189
Reviewed by Chris Fleizach.

Source/WebCore:

Propagate aria-readonly to grid descendants if the property is not
explicitly set on the descendant.

Test: accessibility/gtk/aria-readonly-propagated.html
      Additional test cases also added to accessibility/aria-readonly.html

* accessibility/AccessibilityARIAGridCell.cpp:
(WebCore::AccessibilityARIAGridCell::ariaReadOnlyValue):
* accessibility/AccessibilityARIAGridCell.h:
* accessibility/AccessibilityObject.h:

LayoutTests:

Add new test cases to aria-readonly.html. In addition, because the platform-specific
test expectations seem due to how platforms treat whitespace from the test file, hide
the test cases after the test is run so that there are no extraneous characters and
a single expectations file can be shared.

* accessibility/aria-readonly-expected.txt: Updated.
* accessibility/aria-readonly.html: Updated.
* accessibility/gtk/aria-readonly-propagated-expected.txt: Added.
* accessibility/gtk/aria-readonly-propagated.html: Added.
* platform/mac-elcapitan/accessibility/aria-readonly-expected.txt: Removed.
* platform/mac/accessibility/aria-readonly-expected.txt: Removed.
* platform/win/accessibility/aria-readonly-expected.txt: Removed.


Identifier: 188714@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216425 268f45cc-cd09-0410-ab3c-d52691b4dbfc