AX: Setting aria-orientation="horizontal" on a listbox is being overridden on the Mac https://bugs.webkit.org/show_bug.cgi?id=171821
Reviewed by Chris Fleizach.

Source/WebCore:

The Mac's AccessibilityObject wrapper was unconditionally returning vertical
as the orientation of listbox. Removing this override causes the author-specified
value to be exposed, with the default/implicit orientation continuing to be vertical.

No new tests. We already have coverage for the default/implicit orientation on a
listbox. A new test case was added to aria-orientation.html so that we would have
coverage for aria-orientation="horizontal" on a listbox.

* accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper accessibilityAttributeValue:]):

LayoutTests:

* accessibility/aria-orientation-expected.txt: New test case.
* accessibility/aria-orientation.html: New test case.


Identifier: 188737@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@216452 268f45cc-cd09-0410-ab3c-d52691b4dbfc